third_party/blink/web_tests/http/tests/origin_trials/webexposed/digital-credentials-creation-ot-interfaces.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<!-- Generate token with the command:
generate_token.py https://127.0.0.1:8443 WebIdentityDigitalCredentialsCreation --expire-timestamp=2000000000
-- -->
<meta http-equiv="origin-trial"
  content="A1FTopOFuP+2AlMcgocLTPPugwACDBk9O4pJMFzY7QeRH93cy6XxcilUIKg5GEvR1jFVrh3eqR/B+814yetiVAwAAABueyJvcmlnaW4iOiAiaHR0cHM6Ly8xMjcuMC4wLjE6ODQ0MyIsICJmZWF0dXJlIjogIldlYklkZW50aXR5RGlnaXRhbENyZWRlbnRpYWxzQ3JlYXRpb24iLCAiZXhwaXJ5IjogMjAwMDAwMDAwMH0=" />

<title>Digital Credentials Creation API - Creation of DC is supported</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/origin-trials-helper.js"></script>
<script>
    test(t => {
      // If the API is disabled, this will fail with NotSupportedError.
      promise_rejects_js(t, TypeError, navigator.credentials.create({digital: {requests: []}}));
  }, 'Digital Credentials Issuance with no requests should reject with TypeError.');
</script>
